Day 9 with Big Ints (big enough at least ;) and too many opcodes

+++ intcode/src/io.rs
@@ -1,5 +1,6 @@
 use rustyline::Editor;
 
+#[derive(Debug)]
 pub struct ConsoleIo {
     rl: Editor<()>
 }
@@ -13,7 +14,7 @@ impl ConsoleIo {
 }
 
 pub trait ReadInt {
-    fn read(&mut self) -> Result<i32, Err>;
+    fn read(&mut self) -> Result<i128, Err>;
 }
 
 #[derive(Debug)]
@@ -23,7 +24,7 @@ pub enum Err {
 }
 
 impl ReadInt for ConsoleIo {
-    fn read(&mut self) -> Result<i32, Err> {
+    fn read(&mut self) -> Result<i128, Err> {
         match self.rl.readline("> ").unwrap().trim_end().parse() {
             Ok(n) => Ok(n),
             Err(_) => Err(Err::NaN),
@@ -32,11 +33,11 @@ impl ReadInt for ConsoleIo {
 }
 
 pub trait WriteInt {
-    fn write(&mut self, n: i32);
+    fn write(&mut self, n: i128);
 }
 
 impl WriteInt for ConsoleIo {
-    fn write(&mut self, n: i32) {
+    fn write(&mut self, n: i128) {
         println!("{}", n);
     }
 }
